+++ app/src/test/java/com/sebastianvm/musicplayer/repository/genre/FakeGenreRepository.kt
@@ -5,6 +5,6 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 class FakeGenreRepository(private val genres: List<Genre> = listOf()) : GenreRepository {
-    override fun getGenresCount(): Flow<Int> = flow { emit(genres.size.toLong()) }
+    override fun getGenresCount(): Flow<Int> = flow { emit(genres.size) }
     override fun getGenres(): Flow<List<Genre>> = flow { emit(genres) }
 }
